upd2k25

+++ app/handlers.py
@@ -91,12 +91,17 @@ async def convert_and_send_audio(callback: CallbackQuery, state: FSMContext, url
             await callback.bot.delete_message(chat_id=callback.message.chat.id, message_id=reply_message_id)
         except Exception as e:
             print(f"Error deleting message: {e}")
-    audio_file_path = None
+
     try:
         audio_file_path = await download_youtube_audio(url)
+
+        if not audio_file_path or not os.path.exists(audio_file_path):
+            await callback.message.answer("Ошибка: Не удалось создать аудиофайл.")
+            return
+
         audio_file = FSInputFile(audio_file_path)
         await callback.message.answer_audio(audio_file)
-        os.remove(audio_file_path)
+
         con_answer_id = (await state.get_data()).get('con_answer_id')
         if con_answer_id:
             await callback.bot.edit_message_text(
@@ -104,9 +109,13 @@ async def convert_and_send_audio(callback: CallbackQuery, state: FSMContext, url
                 chat_id=callback.message.chat.id,
                 message_id=con_answer_id
             )
+
     except Exception as e:
         await callback.message.answer(f"Ошибка: Не удалось конвертировать аудио! {str(e)}")
-        os.remove(audio_file_path)
+
+    finally:
+        if 'audio_file_path' in locals() and audio_file_path and os.path.exists(audio_file_path):
+            os.remove(audio_file_path)
 
 
 @router.callback_query(F.data.regexp(r'\d+p'))
